Move test to proper package
