+++ pkg/controllers/machinehealthcheck/BUILD.bazel
@@ -34,7 +34,6 @@ go_test(
     deps = [
         "//pkg/apis/machineremediation/v1alpha1:go_default_library",
         "//pkg/consts:go_default_library",
-        "//pkg/utils/conditions:go_default_library",
         "//pkg/utils/testing:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
