+++ pkg/utils/conditions/BUILD.bazel
@@ -24,6 +24,7 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//pkg/apis/machineremediation/v1alpha1:go_default_library",
+        "//pkg/utils/testing:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
